Rewrite bl-recent-files-pipemenu with safe XML parsing
Partially in response to the issues that were being raised in issue
thread #3, I propose to replace the existing bl-recent-files-pipemenu
implementation with this rewrite, which features:

  * safe XML parsing using xmllint from libxml2-utils and XPATHS
  * no Bashisms, it has been written for /bin/dash
  * cleaner syntax allowing for more trivial adoption and extension
  * configuration by command line switches, no editing needed
  * implements the -h flag for displaying usage information
  * different implementation of the recently-used file's clearing (-c):
    remove the file since it will be automatically recreated

@johnraff, I have modelled the logic after your current implementation.
Since I am not an openbox user though, I would like to ask you for your
opinion on this and also if you could possibly test this. I am going to
post this on the #! forums too which a request for testing/comments.

The script implements two different modi.

In the first modus (without the -H flag), a menu is being generated in
which the respective files are being opened by xdg-open (or any other
user-specified command, see the -o flag).

In the second modus (enabled by setting the -H flag), the files will be
opened by the application from which their recently-used entry
originated; for example, a file saved by Firefox will have Firefox as
the associated application in which it will be opened.